swagger-ui basic configuration
