+++ .gitignore
@@ -25,4 +25,3 @@ yarn-error.log*
 /.husky
 
 src/controllers/temp.js
-src/swagger.js
