+++ src/app.js
@@ -6,61 +6,32 @@ const apiRoutes = require("./routes/index");
 const cookieParser = require("cookie-parser");
 require("./db/connection");
 require("dotenv").config();
+const swaggerOptions = require("./swaggerOptions"); // Import the Swagger options from the separate file
 
-const port = process.env.NODE_LOCAL_PORT || 8080;
+
+const port = process.env.NODE_LOCAL_PORT || 3000;
 
 const app = express();
 
-const swaggerOptions = {
-  definition: {
-    openapi: "3.0.0",
-    info: {
-      title: "Books Express API with Swagger",
-      version: "0.1.0",
-      description:
-        "This is a simple API application made with Express and documented with Swagger",
-      license: {
-        name: "MIT",
-        url: "https://spdx.org/licenses/MIT.html",
-      },
-      contact: {
-        name: "Ceyda",
-        url: "https://re-coded.com",
-        email: "ceyda666esen@gmail.com",
-      },
-    },
-    servers: [
-      {
-        url: "http://localhost:3000",
-      },
-    ],
-  },
-  apis: ["'./controllers/userController.js', './swagger.js'"],
-};
+function attachUserToRequest(req, res, next) {
+  res.locals.user = req.user;
+  next();
+}
 
 const middleware = [
   cookieParser(),
   bodyParser.urlencoded({ extended: false }),
   express.json(),
-  attachUerToRequest,
+  attachUserToRequest,
 ];
 
 middleware.forEach((item) => {
   app.use(item);
 });
 
-const specs = swaggerJsdoc(swaggerOptions);
-function attachUerToRequest(req, res, next) {
-  res.locals.user = req.user;
-  next();
-}
 
 
-app.use(
-  "/api-docs",
-  swaggerUi.serve,
-  swaggerUi.setup(specs, { explorer: true })
-);
+app.use("/api-docs", swaggerUi.serve, swaggerUi.setup(swaggerOptions, { explorer: true }));
 
 app.use("/", apiRoutes);
